chore: separate engine and solid-pixi packages
