+++ apps/client/package.json
@@ -14,8 +14,10 @@
     "@mp/math": "workspace:*",
     "@mp/network": "workspace:*",
     "@mp/pixi": "workspace:*",
+    "@mp/solid-pixi": "workspace:*",
     "@mp/time": "workspace:*",
     "@mp/server": "workspace:*",
+    "@mp/engine": "workspace:*",
     "@mp/state": "workspace:*",
     "@mp/data": "workspace:*",
     "@mp/style": "workspace:*",
