+++ apps/client/src/game/AreaDebugUI.tsx
@@ -10,7 +10,7 @@ import {
 } from "@mp/data";
 import { Graphics } from "@mp/pixi";
 import { createEffect, createSignal, useContext } from "solid-js";
-import { Pixi, EngineContext } from "@mp/pixi/solid";
+import { Pixi, EngineContext } from "@mp/solid-pixi";
 import { myCharacter } from "../state/signals";
 import * as styles from "./AreaDebugUI.css";
