+++ apps/client/src/game/AreaScene.tsx
@@ -1,9 +1,9 @@
 import { type AreaResource } from "@mp/data";
 import { TiledRenderer } from "@mp/tiled-renderer";
 import { useContext, createEffect, Index } from "solid-js";
-import { EngineContext, Pixi } from "@mp/pixi/solid";
 import { createQuery } from "@tanstack/solid-query";
 import { loadTiledMapSpritesheets } from "@mp/tiled-renderer";
+import { EngineContext, Pixi } from "../../../../packages/solid-pixi/src";
 import { api } from "../state/api";
 import { myCharacter } from "../state/signals";
 import { CharacterActor } from "./CharacterActor";
