+++ packages/engine/tsconfig.json
@@ -1,0 +1,1 @@
+{ "extends": ["@mp/tsconfig/vite.json", "@mp/tsconfig/tsc.json"] }
