+++ packages/math/src/index.ts
@@ -2,5 +2,4 @@ export * from "./vector";
 export * from "./clamp";
 export * from "./rect";
 export * from "./size";
-export * from "../../pixi/src/engine/camera";
 export * from "./matrix";
