+++ packages/pixi/package.json
@@ -4,29 +4,12 @@
   "private": true,
   "exports": {
     ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.d.js",
-      "types": "./dist/index.d.ts"
-    },
-    "./solid": {
-      "import": "./dist/solid/index.js",
-      "require": "./dist/solid/index.d.js",
-      "types": "./dist/solid/index.d.ts"
+      "import": "./src/index.ts",
+      "require": "./src/index.ts",
+      "types": "./src/index.ts"
     }
   },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc --build",
-    "dev": "tsc --build --watch --preserveWatchOutput"
-  },
   "dependencies": {
-    "@mp/math": "workspace:*",
-    "@mp/time": "workspace:*",
-    "@mp/state": "workspace:*",
-    "@mp/style": "workspace:*",
-    "solid-js": "1.9.1",
     "pixi.js": "8.3.3"
   },
   "devDependencies": {
