+++ packages/pixi/src/index.ts
@@ -1,3 +1,1 @@
 export * from "pixi.js";
-export * from "./engine/engine";
-export * from "./engine/camera";
