+++ packages/solid-pixi/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@mp/solid-pixi",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.d.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc --build",
+    "dev": "tsc --build --watch --preserveWatchOutput"
+  },
+  "dependencies": {
+    "@mp/pixi": "workspace:*",
+    "@mp/math": "workspace:*",
+    "@mp/engine": "workspace:*",
+    "@mp/style": "workspace:*",
+    "solid-js": "1.9.1"
+  },
+  "devDependencies": {
+    "@mp/tsconfig": "workspace:*"
+  }
+}
