+++ packages/solid-pixi/src/application.css.ts
