+++ packages/solid-pixi/src/index.ts
