+++ packages/tiled-renderer/src/renderer.tsx
@@ -1,7 +1,7 @@
 import type { Layer } from "@mp/tiled-loader";
 import type { Accessor } from "solid-js";
 import { createEffect, createMemo, For } from "solid-js";
-import { ParentContext, Pixi } from "@mp/pixi/solid";
+import { ParentContext, Pixi } from "@mp/solid-pixi";
 import { type JSX } from "solid-js";
 import { recallLayer, LayerViewFactory } from "./layer";
 import type { TiledSpritesheetRecord } from "./spritesheet";
